requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
flask==3.0.2
yfinance==0.2.37
matplotlib==3.8.3
pandas==2.2.1
python-dotenv==1.0.1
pyarrow==15.0.0
requests==2.31.0
requests-oauthlib==1.3.1
pyarrow==15.0.2
pytest==8.3.5
